# openapi.yml
openapi: 3.0.0
info:
  title: Gallery App API
  version: 1.0.0
  description: OpenAPI specification for the CSCC09 Assignment 3 Gallery App backend.
servers:
  - url: http://localhost:3000
paths:
...
  /api/galleries/{galleryId}/images:
    get:
      summary: Get images for a gallery
      parameters:
        - in: path
          name: galleryId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of images
    post:
      summary: Add an image to a gallery
      parameters:
        - in: path
          name: galleryId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                imgFile:
                  type: string
                  format: binary
      responses:
        '201':
          description: Image added
  /api/galleries/{galleryId}/images/{imageId}:
    delete:
      summary: Delete an image from a gallery
      parameters:
        - in: path
          name: galleryId
          required: true
          schema:
            type: integer
        - in: path
          name: imageId
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Image deleted
...